Reader/Runtime: Add Memory and Data support
This adds support for reading Memory and Data sections in wasm binaries,
as well as Memory imports.

Data entries are evaluated much like globals with having to compile and
evaluate the expression of instructions to determine the offset.

This also cleans up the integration test to use a MemberData attribute
to determine the files to run, without having to specify them manually.
Now, any .wat files in the IntegrationTests folder are automatically
run.

Some additional test header operations are added like memory and
expect_call.