Runtime/Compiler: Initialize Table Elements upon module instantiation
This also adds support for the ref.func opcode.

Note that this is a little hacky still. I'd like to clean up
EmitAssembly to have a reusable type for the function holders, and I
don't like the type heirarchy with WasmReferenceType/WasmValueType. This
code also should better align with section 4.5.4 (Execution / Modules /
Instantiation) of the WebAssembly spec. But this is making progress!