Core: Refactor WasmValueType heirarchy to separate types from values
Initially I had made the (in hindsight) strange decision to have wasm
"values" inherit from a type heirarchy that matched the wasm spec for
types. Really, these two concepts (types and values) are separate, and
should not be combined. This refactoring untangles that a bit.

It might not still be the best solution long-term, but at least now the
type heirarchy is less confusing, I think.