Memory: Rename static/dynamic offsets
The WASM spec takes an unusual definition of "operand" to also mean
values that can be on the stack, and not just part of the instruction
itself. This confusion resulted in me previously interpreting the spec
to mean that the "dynamic" offset was the instruction operand while the
"static" offset was implicitly the offset on the stack. This seemed
incorrect, but I assumed it was a spec bug. I found elsewhere in the
spec where they unconventionally define operands to include values
that can be on the stack, meaning that both the static and dynamic
offsets are "operands." Thus, here I've updated the "dynamic" offset to
be the value on the stack which makes the most sense and is most
assuredly what the spec authors intended.