Reader/Compiler: loop and i32.lt_s opcodes
i32.lt_s is really only implemented in here in service of the example
unit test.

Also, as expected in a prior commit, `bool` being the stack type as a
result of equality/less-than/etc opcodes was incorrect. The IL
instructions push an int32 value anyways, so this updates those stack
values to represent reality.