Compiler/Runtime: Support passive data segments, and data count
Turns out wat2wasm will only generate the data-count segment (which
solely contains the number of data segments, for validation) if the
memory.init and/or data.drop opcodes are used. This adds a test for both
of those opcodes, and support for them in the compiler.

In order to use these opcodes, support for passive data segments is
added. Another test is added for multiple data segments, just to ensure
data indexes are working correctly.