Compiler: Add i32.store8 and i32.store16 opcodes
These take the lower 8 or 16 bits and store to a 32-bit memory space.
This tests the functionality by intentionally including some high bits
that would be dropped.

Helpful tip when developing a test case for this: convert a binary value
like 0b1100101010 to int. 0b101010 is 42, so the high two set bits would
be dropped.