Compiler: Add i32.wrap_i64, some test cleanup
This adds the i32.wrap_i64 opcode. Somewhat surprisingly (to this
author), the CIL opcode for this just conv.i4. I guess I had assumed
that this opcode would do a checked conversion, but that must be the ovf
versions of conv.

This also removes some unused memory segments from the reinterpret tests
(which was a copy/paste fail), as well as adds some logging to the test
execution to help confirm that it's doing what you're expecting it to be
doing (instead of silently passing).