Compiler: Add [f32|f64].convert_[i32|i64]_[s|u] opcodes
This uses conv.r4 and conv.r8 in the signed cases, and conv.r.un +
conv.r4/conv.r8 in the unsigned cases. It might not be strictly needed
to do the extra conversion in the unsigned case, but conv.r.un can
seemingly return either single or double precision values depending on
its input, while the WASM opcodes expect a specific type.