Compiler: Fix return type of select opcode, other diagnosic fixes
The select opcode was always pushing i32 onto the type stack (used for
validation) when it should have been pushing the type of its selected
operands. This is now fixed, which allows the Emscripten "hello, world!"
example to compile fully. (It still fails with a "Common Language
Runtime detected an invalid program." exception currently, though.)

In service of diagnosing these issues, I've added a Debug print of which
instruction is being compiled with its operands, and the state of the
current type validation stack.