Compiler: Refactor out Reflection.Emit IL generation
This introduces an IILGenerator abstraction with pulling the
Reflection.Emit implementation out into its own class, so that
WasmCompiler does not depend on a specific IL generator.

This will allow us to experiment with swapping out Reflection.Emit for
Mono.Cecil, as well as unit testing the methods that call the
abstraction without needing to generate a real assembly.

The opcodes that take parameters are strongly typed (unlike with
Reflection.Emit's API), and all opcodes that do not take a parameter are
in an enum. This should help prevent accidental "Common Language Runtime
detected an invalid program" errors in the future. So in a way, this
change helps ensure we're emitting the correct arguments for these
opcodes (even though most of them were covered under tests previously
anyways).

This isn't a perfect abstraction, and it might not survive contact with
the enemy. But it should be a good start towards experimenting with
Mono.Cecil.