Compiler: Move construction of WasmCompiler into CompilationAssembly
This completes the refactoring to remove the dependency on
Reflection.Emit in WasmCompiler. This also narrows the signature of the
WasmCompiler ctor to not require a MethodInfo when it only needs the
method's return type. (Even then, a `bool isVoid` parameter could have
been sufficient, but leaving it as a type for now.)