Compiler: Flatten block/loop expressions for local determination
There was a bug in the determination of if there were any instructions
that needed temporary locals for i.e. args reordering/array-ification,
where it wasn't looking deep into block/loop instructions that may
contain child instructions. This flattens the set of instructions for
the purposes of making this determination, and fixes a bug where the
compiler was emitting a `stloc -1` instruction that the CLR obviously
didn't care much for.