Compiler: Fix ordering of call_indirect args reconstruction
There was a bug in the implementation of call_indirect in the compiler,
where it was popping the function index off the stack _after_ preparing
the args array, when the call index is expected to be at the top of the
stack. This adds a unit test to cover this scenario of using
call_indirect with arguments, as well as some additional out-of-bounds
checking which are useful when debugging these issues.