Compiler: Add Mono.Cecil compiler backend + Emscripten Hello World!
After a bit of bug-fixing in the prior commits, the Mono.Cecil backend
is now fully working with all existing tests. Additionally, this finally
adds the Emscripten "hello, world!" test that the prior commits have
been working up to supporting! Finally, C code that prints to the
console can successfully run via WebAssembly compiled to .NET.

In addition to being an alternative implementation to Reflection.Emit,
the Mono.Cecil backend has the benefit of being able to export its
compiled assembly to disk. This will be done in the future, but was
helpful in debugging the problems with getting the Emscripten example
working.

Because the Mono.Cecil instruction generator validates the opcodes'
operands, it was able to catch the bug with block/loop instructions
fixed in a prior commit. This is much more helpful than the "detected an
invalid program" message that happens if you emit a silly instruction
like `stloc -1` with Reflection.Emit, which does no such validation.