Runtime: Support automatic start function invocation upon instantiation
This by default will now invoke a `start` section-designated function
upon module instantiation if present, or if one is not found, an
exported function named `_start` to match Wasmtime behavior. This can be
disabled with an option flag passed into WasmRuntime (not yet available
in the CLI). With this change, the WasmNet CLI now behaves identically
to Wasmtime for some of the compiled unit tests that have start
functions.