WASI: Fix issue with reading bytes as a string in fd_write
The prior code incorrectly assumed that these byte arrays read from WASM
memory would always end in \0, but this is not true for the "hello,
world!" case when Emscripten is called with -O1. This adds a test for
that example, and better handles null-terminated strings (as well as
those that are not null-terminated).