Compiler: Add `if` opcode support, with `brfalse` CIL compilation
This adds preliminary support for the `if` opcode, which utilizes two
`block`s and the CIL `brfalse` opcode. I'm not 100% convinced that this
is rock solid, but simple tests pass. Still getting an invalid CIL
generation in the -O2 optimized Emscripten "hello, world!" example, but
making progress as this optimized version depended on the `if` opcode.